fix: parallelize session close, document stale ack edge case, extract admin types
1. Parallelize CloseShareSessionsAsync — send close requests to all
   brokers concurrently via Task.WhenAll, consistent with PollAsync.

2. Document the stale pending acks edge case on InvalidShareSessionEpoch
   reset — CommitAsync will re-queue failed acks for retry, so this is
   safe but worth calling out.

3. Extract ShareGroupDescription, ShareGroupMemberDescription,
   ListShareGroupsOptions, ShareGroupOffsetDescription, and
   ShareGroupOffsetAlteration from IAdminClient.cs into dedicated
   ShareGroupTypes.cs for easier navigation.
